[encryption-service] Fall back to old home on failure. Fixes JB#46905
Change nemo's home dir back to /home/nemo every time finish script is
called. Move data back to /home partition only if it's not there already
in case encryption was "cancelled" early. Clean up /tmp. Call preload
ambience in preparation script.

+++ encryption-service/preparation/home-encryption-finish.sh
@@ -1,15 +1,15 @@
 #!/bin/sh
 
 # If encryption finished, remove marker file, otherwise exit
-[ -s /etc/crypttab ] || exit 4
-rm -f /var/lib/sailfish-device-encryption/encrypt-home
+[ -s /etc/crypttab ] && rm -f /var/lib/sailfish-device-encryption/encrypt-home
+
+usermod --home /home/nemo nemo
 
 # Move /home back to /home partition if it is mounted
-if $(mount | grep -q " on /home type"); then
-    usermod --move-home --home /home/nemo nemo
+if $(mount | grep -q " on /home type" && [ ! -e /home/nemo ]); then
+    # /home was wiped, copy stuff back
     mv --target-directory=/home/ /tmp/home/.[!.]* /tmp/home/*
-    add-oneshot --user --late preload-ambience
-else
-    usermod --home /home/nemo nemo
-    exit 5
 fi
+
+# Clean up
+rm -rf /tmp/home/
